AMM-514 : 1097- Moving back to previous page take more time  and image api change

+++ src/main/java/com/iemr/common/identity/data/rmnch/RMNCHBeneficiaryDetailsRmnch.java
@@ -54,7 +54,7 @@ public class RMNCHBeneficiaryDetailsRmnch {
 
 	@Expose
 	@Column(name = "BeneficiaryRegID")
-	private BigInteger BenRegId;
+	private Long BenRegId;
 
 	@Expose
 	@Column(name = "aadhaNo")
