+++ src/main/java/com/iemr/common/identity/repo/rmnch/RMNCHBornBirthDetailsRepo.java
@@ -36,5 +36,5 @@ public interface RMNCHBornBirthDetailsRepo extends CrudRepository<RMNCHBornBirth
 	public RMNCHBornBirthDetails getByIdAndVanID(@Param("vanSerialNo") Long vanSerialNo, @Param("vanID") int vanID);
 
 	@Query(" SELECT t FROM RMNCHBornBirthDetails t WHERE t.BenRegId =:benRegID ")
-	public RMNCHBornBirthDetails getByRegID(@Param("benRegID") BigInteger benRegID);
+	public RMNCHBornBirthDetails getByRegID(@Param("benRegID") Long benRegID);
 }
